requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-appconf==1.0.3
django-avatar==4.1.0
django-bootstrap4==1.1.1
django-braces==1.14.0
django-countries==5.5
django-crispy-forms==1.8.1
django-emoji==2.2.2
django-emoji-picker==0.0.5
django-hitcount==1.3.1
django-imagekit==4.0.2
django-model-utils==4.0.0
django-private-chat==0.2.2
django-taggit==1.2.0
django-tagulous==0.14.1
django-tinymce==2.8.0
django-wysiwyg==0.8.0
djangorestframework==3.11.0
idna==2.8
pilkit==2.0
Pillow==6.2.1
python-dateutil==2.8.1
pytz==2019.3
pyuploadcare==2.6.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
websockets==3.2